test/#353: DTO Validation test를 Controller test로 수정
